@@ -50,10 +50,14 @@ extension OnPressed on Widget {
           right: 0,
           top: 0,
           bottom: 0,
-          child: FlatButton(
-            shape: RoundedRectangleBorder(
-              borderRadius: borderRadius
-            ),
+          child: TextButton(
+              style: ButtonStyle(
+                  shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                      RoundedRectangleBorder(
+                          borderRadius: borderRadius
+                      )
+                  )
+              ),
               onPressed: () {
                 if (onPressed != null) {
                   onPressed();
